requirements.txt
requests~=2.28.1
pandas~=1.2.4
psycopg2-binary==2.9.3
sqlalchemy==1.4.31
polygon